requirements.txt
numpy>=1.24.3
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.36.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# Optional dependencies
chromadb>=0.4.15

# Development dependencies  
pytest>=7.0
pytest-cov>=4.0
pytest-benchmark>=4.0
black>=23.0
isort>=5.0
flake8>=6.0
pre-commit>=3.0
jupyter>=1.0
matplotlib>=3.7.2
seaborn>=0.12.2
